codded entry point
