+++ src/main/java/executor/service/App.java
@@ -1,8 +1,24 @@
 package executor.service;
 
+import executor.service.config.di.Application;
+import executor.service.config.di.ApplicationContext;
+import executor.service.service.ParalleFlowExecutorService;
+
+import java.util.HashMap;
+import java.util.Map;
+
 public class App {
 
     public static void main( String[] args ) {
-        System.out.println("Hello");
+        ApplicationContext context = Application.run(
+                getPackageName(),
+                new HashMap<>(Map.of(ParalleFlowExecutorService.class, ParalleFlowExecutorService.class)));
+        ParalleFlowExecutorService service = context.getObject(ParalleFlowExecutorService.class);
+        service.execute();
+    }
+
+    private static String getPackageName() {
+        Package currentPackage = App.class.getPackage();
+        return currentPackage.getName();
     }
 }
